introduce jws.Verifier
